Embed response
I made the response an embed, and the title is the aliases, color is the same red in the info embed, and the footer shows who it was requested by. it's not much, but it looks nice

Future: I want to add a notation, basic, so if something like  `{.User.Username}` is in the tag content, it'll replace with the username of the person who triggered it.